move BackgroundManager to it's own package outside of the service package
