+++ topl-service/src/main/java/io/matthewnelson/topl_service/service/components/binding/TorServiceBinder.kt
@@ -69,7 +69,7 @@ package io.matthewnelson.topl_service.service.components.binding
 import android.content.Intent
 import android.os.Binder
 import io.matthewnelson.topl_service.service.BaseService
-import io.matthewnelson.topl_service.service.components.BackgroundManager
+import io.matthewnelson.topl_service.lifecycle.BackgroundManager
 import io.matthewnelson.topl_service.util.ServiceConsts.BackgroundPolicy
 import io.matthewnelson.topl_service.util.ServiceConsts.ServiceAction
 import kotlinx.coroutines.Job
